Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -2,22 +2,22 @@
 
 int main(){
     float a,b;
-    float add,sub,mul,div;
+    int add,sub,mul,div;
     char c;
     scanf("%f%f%c",&a,&b,&c);
         if(c=='+'){
            add=a+b;
-           printf("%f",add);}
+           printf("%d",add);}
         else if(c=='-'){
             sub=a-b;
-            printf("%f",sub);
+            printf("%d",sub);
         }
         else if(c=='*'){
             mul=a*b;
-            printf("%f",mul); }
+            printf("%d",mul); }
         else if(c=='/'){
             div=a/b;
-            printf("%f",div);
+            printf("%d",div);
         }
         else{
             printf("error");
